requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfplumber==0.10.3
spacy==3.7.4
sentence-transformers==2.6.1
scikit-learn==1.4.2
numpy==1.26.4
faiss-cpu==1.8.0
requests==2.31.0
pydantic
python-dotenv
google-generativeai

# fastapi
# uvicorn
# pdfplumber
# spacy
# sentence-transformers
# scikit-learn
# numpy
# python-multipart
# requests
